refactor(common): move LCP image observer to a separate file (#46295) (#47082)
This is a small refactoring to extract the LCP image observer to a separate file, similar to the preconnect link checker class location.
PR Close #47082
